+++ aws-sdk-core/lib/aws-sdk-core/xml/parser/engines/libxml.rb
@@ -17,9 +17,18 @@ module Aws
           parser.parse
         end
 
-        def on_start_element_ns(element_name, attributes, *ignored)
+        def on_start_element_ns(element_name, attributes, prefix = nil, uri = nil, ns = {})
+          # libxml-ruby does not provide a mapping from element attribute
+          # names to their qname prefixes. The following code line assumes
+          # that if a attribute ns is defined it applies to all attributes.
+          # This is necessary to support parsing S3 Object ACL Grantees.
+          # qnames are not used by any other AWS attribute. Also, new
+          # services are using JSON, limiting the possible blast radius
+          # of this patch.
+          attr_ns_prefix = ns.keys.first
           @stack.start_element(element_name)
           attributes.each do |attr_name, attr_value|
+            attr_name = "#{attr_ns_prefix}:#{attr_name}" if attr_ns_prefix
             @stack.attr(attr_name, attr_value)
           end
         end